LayoutTests/inspector/file-system-project-expected.txt

Tests file system project.


Running: testFileSystems
Adding first file system.
Adding second file system.

UISourceCode: file:///var/www/localhost/foo.js
UISourceCode is editable: true
UISourceCode is content script: false
Mime type: text/javascript
UISourceCode content: <foo content>

UISourceCode: file:///var/www/bar.js
UISourceCode is editable: true
UISourceCode is content script: false
Mime type: text/javascript
UISourceCode content: <bar content>

UISourceCode: file:///foo/bar/baz.js
UISourceCode is editable: true
UISourceCode is content script: false
Mime type: text/javascript
UISourceCode content: <baz content>

UISourceCode uri to url mappings:
    filesystem:/var/www/localhost/foo.js -> http://localhost/foo.js
    filesystem:/foo/bar/baz.js -> http://www.example.com/baz.js
UISourceCode url to uri mappings:
    http://localhost/foo.js -> filesystem:/var/www/localhost/foo.js
    http://www.example.com/baz.js -> filesystem:/foo/bar/baz.js
Removing second file system.
    number of uiSourceCodes in workspace after removing second file system: 1
Removing first file system.
    number of uiSourceCodes in workspace after removing first file system: 0
